fix MSVC warning

+++ core/utilities/geolocation/engine/plugins/runner/osm/o5mreader/o5mreader.cpp
@@ -293,7 +293,7 @@ O5mreaderIterateRet o5mreader_iterateDataSet(O5mreader* pReader, O5mreaderDatase
 
             int ret = (int)fseek(
                 pReader->f,
-                (pReader->current - ftell(pReader->f)) + pReader->offset,
+                (long)((pReader->current - ftell(pReader->f)) + pReader->offset),
                 SEEK_CUR
             );
